requirements-dev.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev-only
#

-i https://pypi.org/simple
alembic==1.5.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
apache-airflow[google]==1.10.14
apispec[yaml]==1.3.3
appdirs==1.4.4
argcomplete==1.12.2
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
babel==2.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
black==20.8b1
cached-property==1.5.2
cachetools==4.2.1; python_version ~= '3.5'
cattrs==1.4.0; python_version >= '3.7'
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorlog==4.0.2
configparser==3.5.3; python_version >= '2.6'
croniter==0.3.37; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==3.4.7; python_version >= '3.0'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dill==0.3.3; python_version >= '2.6' and python_version != '3.0'
dnspython==2.1.0; python_version >= '3.6'
docutils==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==1.1.2
flake8==3.8.4
flask-admin==1.5.4
flask-appbuilder==2.3.4; python_version >= '3.6'
flask-babel==1.0.0
flask-caching==1.3.3
flask-jwt-extended==3.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
flask-login==0.4.1
flask-openid==1.2.5
flask-sqlalchemy==2.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-swagger==0.2.14
flask-wtf==0.14.3
flask==1.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
funcsigs==1.0.2
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc,grpcgcp]==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-api-python-client==1.12.8
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3; python_version >= '3.6'
google-auth==1.28.0
google-cloud-bigquery-storage==2.3.0
google-cloud-bigquery[bqstorage,pandas]==2.13.1; python_version < '3.10' and python_version >= '3.6'
google-cloud-bigtable==1.7.0
google-cloud-container==1.0.1
google-cloud-core==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.37.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2; python_version >= '3.5'
google-resumable-media==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos[grpc]==1.53.0; python_version >= '3.6'
graphviz==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpc-google-iam-v1==0.12.3
grpcio-gcp==0.2.2
grpcio==1.36.1
gunicorn==20.0.4; python_version >= '3.4'
httplib2==0.19.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-resources==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
iniconfig==1.1.1
iso8601==0.1.14
isort==5.8.0
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
libcst==0.3.17; python_version >= '3.6'
lockfile==0.12.2
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==2.6.11
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1; python_version >= '3.6'
marshmallow==2.21.0
mccabe==0.6.1
mypy-extensions==0.4.3
natsort==7.1.1; python_version >= '3.4'
numpy==1.20.1; python_version >= '3.7'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas-gbq==0.14.1
pandas==1.2.3
pathspec==0.8.1
pendulum==1.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prison==0.1.3
proto-plus==1.18.1; python_version >= '3.6'
protobuf==3.15.6
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyarrow==3.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydata-google-auth==1.1.0
pyflakes==2.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.8.1; python_version >= '3.5'
pyjwt==1.7.1
pyopenssl==20.0.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.17.3; python_version >= '3.5'
pytest==6.2.2
python-daemon==2.3.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==5.4.1
regex==2021.3.17
requests-oauthlib==1.3.0
requests==2.23.0; python_version >= '3.0'
rsa==4.7.2; python_version >= '3.6'
ruamel.yaml.clib==0.2.2; python_version < '3.9' and platform_python_implementation == 'CPython'
ruamel.yaml==0.16.12
setproctitle==1.2.2; python_version >= '3.6'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy-jsonfield==0.9.0; python_version >= '3.5'
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.15
tabulate==0.8.9
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
werkzeug==0.16.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
wtforms==2.3.3
zope.deprecation==4.4.0
